PERA-2916 :: Fix firebase screen log crash
